+++ TerranEditor/res/shaders/Base/Text/TextFragment.glsl
@@ -13,7 +13,7 @@ struct VS_OUT
 layout(location = 0) in VS_OUT fsIn;
 
 // TODO change this to an int
-layout(location = 2) in flat float f_TexIndex;
+layout(location = 2) in flat int f_TexIndex;
 
 layout(location = 0) out vec4 o_Color;
